+++ pkg/analyzer/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Remove deprecated `RecordType.sortedNamedTypes`.
 * Remove `ElementLocation` class; its values are not returned anymore.
 * Remove deprecated `AnalysisContext.analysisOptions`.
+* `LibraryFragment` is not `Annotatable` anymore.
 * Deprecate `element2.dart` library; import `element.dart`.
 * Deprecate `XyzElement2` classes; use `XyzElement` instead.
 * Deprecate `AnalysisError.correction` field; use